server/requirements.txt
# Function dependencies, for example:
# package>=version
Flask
google-cloud-firestore
python-dateutil